Move tests in TagsCloudVisualization_Tests project
