+++ cs/TagsCloudVisualization_Tests/Program.cs
@@ -1,0 +1,9 @@
+﻿namespace TagsCloudVisualization_Tests
+{
+    public class Program
+    {
+        static void Main()
+        {
+        }
+    }
+}
