+++ cs/tdd.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Samples", "Samples\Samples.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualization", "TagsCloudVisualization\TagsCloudVisualization.csproj", "{14C77268-0FF9-479D-9D3E-3C5163FFD09A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TagsCloudVisualization_Tests", "TagsCloudVisualization_Tests\TagsCloudVisualization_Tests.csproj", "{2397AFC4-6035-44BC-B000-D09F4B621D9B}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -27,6 +29,10 @@ Global
 		{14C77268-0FF9-479D-9D3E-3C5163FFD09A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{14C77268-0FF9-479D-9D3E-3C5163FFD09A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{14C77268-0FF9-479D-9D3E-3C5163FFD09A}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2397AFC4-6035-44BC-B000-D09F4B621D9B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2397AFC4-6035-44BC-B000-D09F4B621D9B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2397AFC4-6035-44BC-B000-D09F4B621D9B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2397AFC4-6035-44BC-B000-D09F4B621D9B}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
